[refactor]: requested changes
