+++ lib/pages/mortgage/components/mortgage_stock_item.dart
