+++ lib/pages/mortgage/retrieve_page.dart
@@ -1,6 +1,6 @@
 import 'package:dalal_street_client/blocs/mortgage/mortgage_details/mortgage_details_cubit.dart';
 import 'package:dalal_street_client/config/get_it.dart';
-import 'package:dalal_street_client/pages/mortgage/components/stock_retrieve_item.dart';
+import 'package:dalal_street_client/pages/mortgage/components/retrieve_stock_item.dart';
 import 'package:dalal_street_client/proto_build/models/Stock.pb.dart';
 import 'package:dalal_street_client/streams/global_streams.dart';
 import 'package:flutter/material.dart';
@@ -46,7 +46,7 @@ class _RetrievePageState extends State<RetrievePage> {
             itemBuilder: (context, index) {
               int stockId = state.mortgageDetails[index].stockId;
               return RetrieveStockItem(
-                  company: mapOfStocks[stockId] ?? Stock(),
+                  company: mapOfStocks[stockId]!,
                   mortgageDetail: state.mortgageDetails[index]);
             },
             separatorBuilder: (BuildContext context, int index) {
